+++ Realm/src/main/java/org/jmangos/realm/network/packet/wow/client/CMSG_CHAR_CREATE.java
@@ -91,7 +91,7 @@ public class CMSG_CHAR_CREATE extends AbstractWoWClientPacket {
         
         final CharacterData charData = new CharacterData();
         // Set account id
-        charData.setAccount((int) getAccountInfo().getId());
+        charData.setAccount(getAccountInfo().getId());
         
         // Set name
         charData.setName(this.charName);
